Organize client project
