+++ React/shop/src/Components/ThreeDPage/ThreeDPage.js
@@ -1,15 +1,19 @@
-import NavigationBar from './NavigationBar';
-import "./AudioPage.css";
+import NavigationBar from '../NavigationBar/NavigationBar';
+import "./ThreeDPage.css";
 import { Component, useEffect, useState } from 'react';
-import SubCategoriesFilter from './SubCategoriesFilter';
-import ProductsControl from './ProductsControl';
+import SubCategoriesFilter from '../SubCategoriesFilter/SubCategoriesFilter';
+import ProductsControl from '../ProductsControl/ProductsControl';
 
-class AudioPage extends Component {
+class ThreeDPage extends Component {
   constructor(props){
     super(props);
   }
 
-  
+   componentDidMount() {
+    if(this.props.products.length > 0)
+      console.log("3d load")
+   }
+ 
 
     render (){    
       return(
@@ -34,4 +38,4 @@ class AudioPage extends Component {
     }
   }
   
-  export default AudioPage;
+  export default ThreeDPage;
