+++ React/shop/src/Components/TwoDPage/TwoDPage.css
@@ -1,0 +1,21 @@
+.main-grid{
+    display: grid;
+    grid-template-columns: 250px 30px auto;
+    }
+  
+    .products-grid{
+      grid-column-start: 3;
+      grid-column-end: 4;
+      grid-row-start: 1;
+  width: 100%;
+  float: right;
+    }
+  
+    .controls-grid{
+      grid-column-start: 1;
+      grid-column-end: 2;
+      grid-row-start: 1;
+  width: 100%;
+  float: left;
+    }
+
