+++ React/shop/src/Components/VFXPage/VFXPage.css
